requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain-groq
duckduckgo_search==5.3.1b1
pymupdf
arxiv
wikipedia
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
langchain
langgraph
langchain-core
langchain-community
python-dotenv
langchain-openai
langchain-groq
arxiv
wikipedia
langgraph-cli[inmem]
faiss-cpu
langextract
poml
langchain_mcp_adapters
azure-storage-blob
azure-storage-queue
azure-storage-file-share
azure-cosmos
azure-keyvault-secrets
azure-keyvault-keys
azure-keyvault-certificates
azure-identity
bcrypt
streamlit
streamlit-cookies-manager
azure-eventhub
azure-eventhub-checkpointstoreblob-aio
streamlit_autorefresh
cookie_manager